Added in specular lighting

+++ Light.cpp
@@ -21,8 +21,8 @@ Light::~Light()
 {
 }
 
-void Light::UseLight(GLfloat ambientIntensityLocation, GLfloat ambientColorLocation,
-                     GLfloat diffuseIntensityLocation, GLfloat directionLocation)
+void Light::UseLight(GLuint ambientIntensityLocation, GLuint ambientColorLocation,
+                     GLuint diffuseIntensityLocation, GLuint directionLocation)
 {
     glUniform3f(ambientColorLocation    , color.x, color.y, color.z);
     glUniform1f(ambientIntensityLocation, ambientIntensity);
